Update panels.mdx
Feel free to toss this if I am wrong, but pretty sure iOS or Panels one made it so that you can use HTML. I’ve been using it for a long time, and found references on the Discord of at least one other person getting it to work with HTML only. 

I know that it is kind of a barrier for people who don’t know how to set up a certificate, but obviously is more secure if they do, so I don’t know if you might want to keep it, or change it to a recommendation.

I’m not sure what would be best, but I trust you, Diesel, to know what to do, just thought I would point it out.